Renames spatial_vector.h to spatial_velocity.h
