+++ drake/multibody/multibody_tree/math/spatial_velocity.h
